Refactor: update import path for ImageCropScreen components
